// src/common/constants/error-messages.enum.ts
export enum ErrorMessages {
  SERVICE_ERROR = '_SERVICE_ERROR',
  MAIL_SEND_FAIL = '메일 발송에 실패했습니다.',
  NOT_FOUND_REQUEST = '요청을 찾을 수 없습니다.',
  INVALID_REQUEST_PARAMETER = '요청 정보가 유효하지 않습니다.',
  INTERNAL_SERVER_ERROR = '알 수 없는 오류가 발생했습니다.',
  NOT_FOUND_USER = '사용자 정보를 찾을 수 없습니다.',
  WITHDRAWAL_USER = '탈퇴한 사용자입니다.',
  ACCESS_DENIED = '접근 권한이 없습니다.',
  EXISTED_EMAIL = '이미 존재하는 이메일입니다.',
  EXISTED_PHONE = '이미 존재하는 전화번호입니다.',
  DUP = '중복된 데이터입니다.',
  NOT_FOUND_DATA = '데이터를 찾을 수 없습니다.',
  INVALID_EMAIL_CODE = '이메일 인증 코드가 유효하지 않습니다.',
  EXPIRED_EMAIL_CODE = '이메일 인증 코드가 만료되었습니다.',
  INVALID_TOKEN = 'INVALID_TOKEN',
  NOT_FOUND_TOKEN = 'NOT_FOUND_TOKEN',
  EXPIRED_TOKEN = 'EXPIRED_TOKEN'
}